\name{getDMgenes}
\alias{getDMgenes}
\title{getDMgenes}
\description{getDMgenes}
\usage{
getDMgenes(dmprobes, background)
}
\arguments{
  \item{dmprobes}{Character vector of Illumina 450k IDs (e.g. "cg00000108"") constituting subset that is differentially methylated}
  \item{background}{Object of type "background" generated by \code{\link{create450kbg}}}
}
\value{Character vector of gene symbols}
\author{Tim J. Peters <Tim.Peters@csiro.au>}
\examples{
\dontrun{
bg <- getDMgenes(rownames(my.betas), groups=c("TSS1500", "TSS200"))
dmgenes <- getDMgenes(my.dmprobes, bg)
#And then goseq:
runGoSeq(dmgenes, all.genes=names(bg$background), onto="GO:BP", genome="hg19", ID="geneSymbol", bias.data=bg$background)
}
}